requirements.txt
Flask
pyserial
requests
paho-mqtt
tornado